Remove some pesky message boxes; implement shape styles for the rectangle tool!
Message boxes are annoying, especially when 50 are spawned on top of
each other because you moved your mouse.
It's much better to have non-functioning options that you can change.
But you know what's even better? Working options. So I made this work
with the rectangle tool easily enough.
This also might fix an issue with the corners of rectangles being of
different color from the rest of the stroke.