Fix minor issue with double-click docking
Before, it didn't save where it docked to when you dragged a window onto
a side, so when you dragged a window onto a side, then dragged the
component back out into a window, then double-clicked on the title bar,
it would go to the wrong position. So yeah, that's fixed.