Improve title
Maybe this will make it show up nicer in Google?
Currently Google is transforming "untitled - Paint" to "Paint: untitled" for the title of the page; maybe this would make it more confident in using the original title, since it lines up with the domain name, and is more than one word? idk