Select multiple tools haha this sucks
It's not even bad in a fun way, because the tools share state and basically one wins, for the most part.

You can combine airbrush and brush which is nice... altho you have to select airbrush second for it to be continuous over time.

If you select select as the last thing you can use it with other tools, and then with Ctrl+I to invert, you can do something like this: https://i.imgur.com/DXbc2GT.png