Prevent layout shift while loading
Fix dynamic cursor offset if you 1. load a new empty session (remove the whole hash from the URL), 2. switch to the brush or eraser tool, 3. see the cursor offset on the canvas

Fix flash of UI where there's no menu bar while loading, that caused the shift in UI that caused the cursor offset to be wrong.