Preserve current branch as redos when navigating history
For instance if you cancel a brush stroke and then cancel another, and redo, it'll now go to the newer brush stroke.
It doesn't just go to the latest tho, if you jump to the first canceled brush stroke in the history tree,
undo and redo and you'll stay with the older brush stroke, staying on that branch of the tree.