Show a specific error for URL not pointing to image
This also makes it not try the WayBack Machine if it already loaded the file and just couldn't decode it as an image - which sounds good, except that websites sometimes redirect swathes of URLs to a new address, no longer relating to the image. In that case falling back to the WayBack Machine could still be useful.