Speed up tool tests
Shaves a few seconds off. Kinda disappointing, I was expecting it to be more of an improvement.
Is it worth the complexity? Well, I can always go back to doing visit() before each test, if resetting the state gets more complicated.
(As long as I don't introduce actual dependency on the state from earlier tests.)