Detect more collisions (and change outcome)
This also changes the localization of "Colors" from "색" to "색상", perhaps among other things. It's hard to view the diff, and I don't know enough about Korean to decide which is better. Probably it should be different in different places, since there are multiple translations, and that will be a separate feature to introduce.