Align radio buttons in Image Attributes
This should help with automated visual tests where the radio buttons were for some reason treated as stacked vertically when running headless, and when running interactively (only way to debug), they would show as expected (a heisenbug).
And of course it brings it closer to MS Paint, and looks nicer.

...NOPE, doesn't help.