Update loose linting
It's insane to maintain this list of var names, to reign in the hundreds of errors that rain down just so wading thru is less of a pain.
There's still far much more chaff then there's grain.

This is obviously not a good linting situation, but it's what I've got right now, and it found a few bugs.
Or, I *managed to find* a few bugs with it, let's not give it too much credit.