Don't prevent default of some controls when refocusing a window
It was like the macOS paradigm before, in that some controls would get triggered when clicking to focus a window and some wouldn't — except with no consideration for the consequences of triggering those controls. For instance, focusing a text field was prevented, while clicking buttons was not!
It could even give you a false sense of security if you noticed that it focused a window without also triggering a control once, and then you go and accidentally trigger some action and have no idea what you did, so it was really the worst of both worlds.
Now it focuses and triggers all the time, as is the Windows paradigm (among other operating systems).

This fix is also a simplification :)