Navigate history view with arrow keys (distinct from Undo/Redo)
How is moving up and down the history view distinct from undo/redo?
Undo (Ctrl+Z) and Redo (Ctrl+Shift+Z) navigate along one branch of history.
When you draw something, undo and draw it another way, you can no longer redo,
because that is the end of the LINEAR path of history.
But history in JS Paint is actually NON-LINEAR: if you use the history view,
you can get to any previous state of the document.
It's ordered chronologically, so starting from the top and going down,
it'll show you the whole of your process of creating a piece of art,
with all the mistakes and attempts and ideas explored along the way.
With arrow key control, you can navigate history in an animated way,
by holding down Down or Up to repeat the key.