Set default dev server port
Keep projects on disinct ports so that 1. the local storage isn't shared with other projects, and I can freely clear it on one project without affecting another, and 2. so it doesn't sometimes get a random free port which then has local storage that is TOO separate, and drawings there can be lost too easily :)

I'd do port 1998 but I think I'll reserve that for working on 98.js.org