Move localization script to a file
There's no real reason for it to be inline, or before the library scripts.
I want it to load faster, sure, but I should set up a build step if I want to improve that.
Oh, there's a document.write(), but it seems to work? I think it just needs to be loaded synchronously.