Improve pinch zooming
I remembered I had implemented pinch zooming fairly nicely in another project: https://github.com/1j01/janitorial-android

It checks for a threshold of distance change before zooming in or out, which has the nice effect that there's no near-boundary pinch levels/distances where it can oscillate more easily; as soon as you cross a boundary to a new zoom level, you have to go the whole threshold back, which makes it much more stable.

It's also flexible in that I was able to define a set of zoom levels, which don't have to be evenly spaced, although it'd be nice to adjust the threshold between them too, maybe just to match mathematically in a logarithmic way, but potentially I could just include thresholds along with the zoom levels list, that'd be pretty simple to do if I wanted fine grained control.
I haven't included a list of zoom levels here, but it's an option.