Improve safety when modifying the canvas while the session is loading
This is not ideal. There are surely race conditions where it can still end up with the document overwritten, and if you hit Cancel after saying Save, ditto.
Ideally it should not be able to overwrite the session while loading in the first place!
I guess that wouldn't be too hard...

This also improves the dialog around saving without a file extension, or with the wrong file extension, which is also part of improving the safety (in this less than ideal scheme).