Separate labels from radios, to support 98.css's styling
It needs label to be a sibling of the radio, to do the styling of the radio buttons using a pseudo element on the label.
Unfortunately, it doesn't just not style radios that aren't in the right structure, it makes them invisible. I think this could be fixed by resetting styles for `label input[type=radio]` (todo)