Work around system overlays obscuring menu bar and tools on iPad
In fullscreen, the iPad shows the system bar as an overlay at the top, making the top of the page inaccessible.
It also adds a floating exit fullscreen (X) button.

This workaround shifts elements around to avoid these overlays.
It adds a spacer element to the top of the page to account for the system bar,
and styles it thematically as a titlebar. It looks great (very legit and intentional), at least on this one model of iPad,
but will probably break on other models. It's a very fragile thing, working around system overlays.