Prevent strange behavior when clicking on custom zoom field
The field itself should be focused when clicking on the field directly,
and the caret should be placed where you click.

You know, I thought of this solution quickly, and GitHub Copilot
autocompleted it, but I'm having trouble pinning down what the problem
actually was.

Why does focusing this input programmatically lead to the input not
being focused ultimately after the click? At any rate, letting it focus
naturally works, and is clearly superior.

Note: This buggy behavior was showing up in Firefox.