Fix opening files with drag and drop on Firefox
When dragging in a file to open in Firefox, it showed an inappropriate
message "File not found", with the expandable error details showing:

    TypeError: item.getAsFileSystemHandle is not a function

That's a missing API, not a missing file.

Another way to handle this would be to detect `getAsFileSystemHandle`
support earlier in order to reuse the `dt.files` code path below.
It might be cleaner.