Add more gradients/coloring to outlines (fill/pencil/brush/airbrush)
I also colored the cap of the airbrush white, but I haven't aligned it to the pixel grid to make sure it shows up at 1x.

I added a separate shape for the ferrule of the pencil (eraser fastener), since previously there were only shapes for two highlight bands of the ferrule's corrugation, which strokes didn't extend to the edge of the pencil, and now it's to be a separate stroke color from the eraser and shaft. I also duplicated the shape of the pencil tip to keep it black while changing the outline color.
In less words, I had to restructure things a bit.