Add translucent white outline around selection marquee for contrast
Unlike in the MS Paint reference version (Windows Vista), I'm allowing the outline to extend outside the yellow "canvas". I would have to go out of my way to make it not apply outside the canvas, and it seems useful there too, and likely only didn't in Windows Vista because of binary transparency.
That said, it can't extend outside the bounds of the SVG, so the very rightmost part of the outline is cut off.