Fix error when changing eraser size with keyboard after clicking a color
Fixes https://github.com/1j01/jspaint/issues/334

When pressing numpad +/- to change the size of the eraser,
it got an error, if previously the canvas was hovered (but the eraser
hadn't been used on it yet), and a color was selected in the palette
(but the canvas wasn't interacted with after selecting the color.)

Either:

    TypeError: pointer_previous is undefined
    
    paint@https://jspaint.app/src/tools.js:366:19
    tool_go@https://jspaint.app/src/app.js:1460:17
    @https://jspaint.app/src/app.js:1044:13
    @https://jspaint.app/src/app.js:1043:20
    dispatch@https://jspaint.app/lib/jquery-3.4.1.min.js:2:42571

Or, if the canvas was interacted with before switching to the Eraser:

    TypeError: this.mask_canvas is null
    
    paint_iteration@https://jspaint.app/src/tools.js:377:5
    paint/<@https://jspaint.app/src/tools.js:367:10
    bresenham_line@https://jspaint.app/src/image-manipulation.js:232:11
    paint@https://jspaint.app/src/tools.js:366:18
    tool_go@https://jspaint.app/src/app.js:1460:17
    @https://jspaint.app/src/app.js:1044:13
    @https://jspaint.app/src/app.js:1043:20
    dispatch@https://jspaint.app/lib/jquery-3.4.1.min.js:2:42571

The code that sets `button` and `ctrl` was introduced in https://github.com/1j01/jspaint/commit/9ad2626713cc6756c93fc927d901c758c68dc304#diff-e07d531ac040ce3f40e0ce632ac2a059d7cd60f20e61f78268ac3be015b3b28fR1353-R1354
back before the Edit Colors dialog was implemented, when first adding support for custom colors, via the browser's native color picker (`<input type="color">`).
With the Edit Colors dialog implemented, the state of which color selection (primary, secondary, or tertiary) to be updated is stored in the scope of the function `show_edit_colors_window`, as `color_selection_slot_to_edit`.
I audited the codebase for all usages of `button` and `ctrl` and I don't think there's any other reason to set these globals when clicking a color button.