Rename Modern Dark theme's stylesheet/ID
This theme is new (as yet unreleased), so I don't need to worry about users with a stored theme preference for this theme.
As for "modern.css", it's used by both the Modern Light and Modern Dark themes, so there's a justification for its name.
As for "dark.css", it ought to be "classic-dark.css" to match, and then I would want to create a tiny translation to propagate the preference.