Update canvas CSS width/height in apply_image_transformation
This was really confusing why my new `repack_spritesheet` function wasn't updating the display size of the canvas, even though it updated the handles, and even though the canvas rotating action updated the canvas properly despite using the same `apply_image_transformation` function and not apparently calling anything to update the canvas display size.
It took me a while to figure out it was updating it OUTSIDE of `rotate`!

I searched the codebase for /"[a-z ]*update[a-z ]*"/g, and didn't find any "update" event handlers that looked like they would do it, and I had to use DOM Mutation Breakpoints to understand what was changing the style attribute.

Also, I'm casting some shade on the other "resize" event triggers.
Some things just fall into a bad pattern of "update the stuff please" vaguery.
It's sort of related to how cache invalidation is the root of all evil.