Add inset shadows to tool options area, to make it feel more physical
The shadows that are part of the border-image don't go on top of the
contents of the element, but I can somewhat mimic the shadows,
using box-shadow on a pseudo element.

This better integrates the contents in the imagined physical scene,
but it does increase the visual weight of the shadows,
where it overlaps the shadows baked into the border-image.