Adjust vector version of Classic Dark tool icons
This updates the Classic Dark vector icons to match the style I came up with in the raster version to improve contrast.

I used the `src/repack-spritesheet.js` script to create the spaced out PNGs to use as a reference in Inkscape.
I may also use them in the future to avoid bleeding / cut-off icons.

(For some reason Inkscape changed all the XML whitespace while ironically adding `xml:space="preserve"`,
but I was able to remove these changes by removing `xml:space="preserve"`, and opening and re-saving the file, and then amending this commit for a clean diff.)