Update screenshot baselines
Note the "visual tests -- modern theme eye gaze mode.snap.png" image,
where the entire icon set is changed, and it wasn't detected as a change
with a threshold of 5% (and the icons are HUGE!)

Also, I'm accepting some negative changes, in the tool options area,
which I see only in the tests, not in Chrome or Firefox or the Electron
app. The magnification levels and transparent/opaque mode icons appear
cut off.

I'm also accepting some barely visible changes at the top of the
toolbox component, where the border has moved down a pixel, but I'm not
accepting some changes that were completely imperceptible, despite being
detected as changes.