Fix clicking radio buttons directly, in Classic/Winter/Occult themes
I always tend to click on the radio option labels, as it's a larger
click target, easier to reach with the mouse, so I missed that clicking
the radio button (circular part) was broken.

98.css hides the real `<input type="radio">` and replaces it with a
pseudo element on the `<label>` element to visually replace it.
Normally it's clickable in the same way that the `<label>` is clickable,
since the pseudo element is logically part of the `<label>`.
However, I have an overly broad selector for pseudo elements, which was
making them unclickable.
I've added an exception to this over-reaching blanket rule for now,
but it should be reigned in properly in the future.

Fixes https://github.com/1j01/jspaint/issues/311
Fixes https://github.com/1j01/jspaint/issues/312