Add tests for Custom Zoom window, and Modern Dark and Bubblegum themes
The new baselines for the Custom Zoom tests incorrectly represent
what real browsers do, although more explicit layout styles could
help for robustness, and for the Bubblegum theme, which actually looks
more worse in reality, unlike the other themes, which look better.

Also, the Modern Dark theme uses `color-scheme: dark` to let the browser
style the buttons and other controls to fit in with a dark theme,
but it's apparently not supported in this version of Electron.