Align emoji in Extras menu
Not all emoji show as emoji on all platforms.
For example, "ℹ️" and "↕️" don't show as emoji for me on Chrome, but
they do on Firefox. (In Sublime Merge, only "ℹ️" shows as an emoji.)

Now that the emoji are pseudo-elements, they can be aligned nicely,
regardless of whether they show up as emoji.