Mark foldable regions in app.js
`#region` and `#endregion` are special comments that can be collapsed in
VS Code, and some other editors, at least with plugins.

By identifying chunks of related code, this is a decent step toward
breaking this huge file up into smaller modules.

By making chunks of code foldable, this also affords some of the same
benefits in the short term, such as being able to drag the scrollbar
with the mouse with reasonable ease and precision.

You can collapse regions with the "Fold All Regions" command in VS Code,
or individually by clicking in the gutter.