Fix GIF not animating in Firefox?
It's weird; Firefox was able to animate the GIF, when loaded in a
separate tab, but in the menu, it wasn't animating.

(I was a little worried about "updated.gif" since it displays stretched
in Windows 11's File Explorer, but it seems fine in practice, whereas
"new2.gif" having problems.)

To fix the problem, I ran the GIF through https://ezgif.com/optimize

However, while this seemed to fix it, checking out the version before
this commit, it seems to be working fine. So I can't say I fixed it,
or that I understand it, or that I can reproduce it. If anyone asks,
I was just shaving 3% of the bytes off of this file as an optimization.