Wait for images to load in visual tests, instead of blacking them out
This reveals that the update indicator (which I'm still blacking out)
in the About window is obscured by the app icon in the Bubblegum theme.

Aside from that, it now tests that the correct icon is shown, and
it's just more satisfying to see the icons, plus it helps with
identifying screenshots at a glance.