Fix error when re-opening closed Electron window on macOS
When the window is closed on macOS, the application continues running,
and handles re-opening the window if you activate the app.
It got an error in the main process due to re-registering event handlers
for each window opened. These are global IPC event handlers, not
specific to a window, so should only be registered once.
This error message showed while opening a new window,
although the app was still able to load:

	Uncaught Exception:
	Error: Attempted to register a second handler for 'show-save-dialog'
		at IpcMainImpl.handle (node:electron/js2c/browser_init:189:337)
		at createWindow (/private/var/folders/s7/w29zcnr9607_rb9fh0sxplt00000gp/T/AppTranslocation/E2038B46-C3E7-47D6-9C4E-08BE5EBFFA7B/d/jspaint.app/Contents/Resources/app/src/electron-main.js:163:10)
		at App.<anonymous> (/private/var/folders/s7/w29zcnr9607_rb9fh0sxplt00000gp/T/AppTranslocation/E2038B46-C3E7-47D6-9C4E-08BE5EBFFA7B/d/jspaint.app/Contents/Resources/app/src/electron-main.js:284:3)
		at App.emit (node:events:526:28)