Handle window creation in open-file just like second-instance
This changes the error that occurs on macOS when trying to start the app
with an open file, from:
  TypeError: Cannot read properties of undefined (reading 'webContents')
To:
  Error: Cannot create BrowserWindow before app is ready

This applies to Open With > JS Paint in Finder, dragging a file onto
JS Paint.app in Finder, and dragging a file onto the app's dock icon,
but only when the app is completely quit. (The app has to be pinned to
the dock to test the last case.)

If the app is running, with or without any windows open, all three ways
to open a file work fine, and this is unaffected.