Handle "Open" links in File > Manage Storage dialog in Electron app
Before, the "Open" links were set to open in a new tab, which was
blocked in the Electron app, as it's designed only as a single-window
app for now.

Now, you can open documents from Manage Storage, in the Electron app,
and they will open in the same window.

However, if you edit the document before clicking an "Open" link,
it triggers a dialog message which is not appropriate to the scenario:

    You've modified the document while an existing document was loading.

    Save the new document?

(This message also inappropriately appears outside of Electron,
if you type a #load: URL into the address bar such as
  http://127.0.0.1:1999/#load:https://i.imgur.com/M5zcPuk.jpeg
after editing the document.)

It's a very confusing message, but at least you can open documents
from Manage Storage now.