Increase resolution of macOS app icon
This drops 48px (even though it has special pixel treatment from MS),
and adds 128px and 512px (from my SVG version).

The 48px version was ignored by iconutil; it's not listed as a size
recommended by Apple, and the utility probably ignores sizes outside
that list, with an exception for the highest resolution.