VS Code: fix inaccurate task name
This task is actually cross-platform and runs the app un-packaged.
It's taken from here: https://www.electronjs.org/docs/latest/tutorial/debugging-vscode
I had renamed it because it was too general (JS Paint isn't only an
Electron app, so I wanted to include "Electron" in the name),
but while I was at it I made it too specific as to what I wanted.

Debugging the packaged app on macOS was my GOAL when I added this task,
not what I achieved with it. (I ended up using console logs to debug,
but left this task in since it seems useful.)