Convert theme.js to ESM
I'm hoping to progressively convert the codebase to ES Modules,
using globals temporarily until all dependent code is converted
to modules, and then switching to using imports, for a given dependency.

Adding `type="module"` makes the script load asynchronously, which means
that dependent code must be deferred in order to access the temporary
globals.

This could affect the flash of unstyled content (FOUC), which makes it
both a good thing to try out first, and potentially something to do last
or not at all, but I'm not noticing a difference before and after this
change. There is some FOUC, but it doesn't appear worsened.