Silence eslint warning by adding @this comment
src/helpers.js
  62:19  error  Unexpected 'this'  no-invalid-this

I tested the value of `this` and it works with debounced methods.

In general it's often good to change your code in ways other than
adding an ignore comment if you can, but I don't really feel like this
counts, as it's just adding a different sort of comment.

Sometimes I just like to do things "fancier", I'm not gonna lie.
I mean look at the TAU constant definition at the top of this file.

At any rate, this brings it back to zero eslint errors,
with the caveat that lots of rules are disabled.