Fix 2 failing Cypress tests
  1) visual tests
       modern light theme -- edit colors dialog (expanded):
     CypressError: Timed out after `4000ms`. The `done()` callback was never invoked!
      at cypressErr (http://localhost:11822/__cypress/runner/cypress_runner.js:158025:16)
      at Object.errByPath (http://localhost:11822/__cypress/runner/cypress_runner.js:158080:10)
      at http://localhost:11822/__cypress/runner/cypress_runner.js:159537:27

  2) visual tests
       winter theme -- edit colors dialog (expanded):
     CypressError: Timed out after `4000ms`. The `done()` callback was never invoked!
      at cypressErr (http://localhost:11822/__cypress/runner/cypress_runner.js:158025:16)
      at Object.errByPath (http://localhost:11822/__cypress/runner/cypress_runner.js:158080:10)
      at http://localhost:11822/__cypress/runner/cypress_runner.js:159537:27

This broke in ce2e1e6f0 when installing typescript,
as determined by git bisect.

Indeed, uninstalling typescript fixes these errors,
and reinstalling it breaks it.
However that's not a very good solution, now is it!

I can't follow the links in the stack traces to try to understand
what is happening here, but I found a way to fix it by looking at the
structural differences between the failing "(expanded)" tests and
the non-failing non-expanded edit colors dialog tests.