Resolve lint in repack-spritesheet.js with a relative URL hack
Add a ".." to the imports so that the paths are understood by
typescript. The script is meant to be pasted into the browser console,
and thus run relative to the page, rather than the file on disk, but
the imports can still actually work in the browser like this, assuming
that the page is served at the root of the web server, since ".." at the
root is ignored.

It looks silly since the paths start with "../src", while the files are
within `src`, making it redundant, but this allows it to target both
environments simultaneously. Isn't that clever?

For more "cleverness" of this sort, just look at "dev:start-server:NOTE"
in `package.json`, regarding an octuple backlash.

I bumped the ecmascript version numbers since I'm using top-level await.