Struggle to make RESTSession more robust
I'm trying to make it so you never lose changes due to your own changes
being synced to the server, but it's still occasionally dropping changes
as soon as you make them. The fact that the changes are reverted as soon
as you make them makes me thing introducing some artificial latencies
may help to debug this, make it occur more often.

I think I'm going to need a more formal testing strategy, or
diagrammatic visualization of the possible timings, or both,
in order to solve this.

However, this architecture is never going to be robust with multiple
users editing at the same time. So, I could move on to more deeply
rearchitecting the multiplayer/tools/history systems.

I don't know if this commit helps.

Another idea I had is to store an edit count along with the image,
and have the highest edit count always win, in order to minimize lost
work, with edits as an approximation of work. This may improve the
multiple-user experience while solving the single-user case completely.
Perhaps. If it's done right.
One finer detail: a tie breaker may be needed, such as a UTC timestamp,
or, for "fairness" (while we're accepting some loss will occur),
a combination of the user id and edit count such that it cycles which
user's edit will win (since even a UTC timestamp could be affected by
clock skew and make one user's edits usually win).