Allow overriding online session implementation for development
The RESTSession requires a server. I don't have hosting set up for it,
yet, but the Discord Activity requires a server anyways.
The FirebaseSession uses Google's servers, and although the quota limits
mean that it's almost never available, it at least in principle is
connected to server/database hosting, and so I don't want to switch
to an implementation that doesn't have hosting set up yet, for the
general case. (Although perhaps there's a case for making a clean break
from Firebase, or in other words, disabling the feature, since it's
been so broken for so long.)