Fix the last remaining TypeScript error
This commit fixes the last remaining type checker error:

    > jspaint@1.0.0 typecheck
    > tsc --noEmit -p jsconfig.json
    
    src/app-state.js:8:1 - error TS6200: Definitions of the following identifiers conflict with those in another file: make_canvas, get_tool_by_id, make_monochrome_palette, make_history_node, TOOL_PENCIL, brush_shape, brush_size, eraser_size, airbrush_size, pencil_size, stroke_size, default_brush_shape, default_brush_size, default_eraser_size, default_airbrush_size, default_pencil_size, default_stroke_size, tool_transparent_mode, stroke_color, fill_color, stroke_color_k, fill_color_k, selected_tool, selected_tools, return_to_tools, selected_colors, selection, textbox, helper_layer, $thumbnail_window, thumbnail_canvas, show_grid, show_thumbnail, text_tool_font, root_history_node, current_history_node, history_node_to_cancel_to, undos, redos, file_name, system_file_handle, saved, pointer, pointer_start, pointer_previous, pointer_active, pointer_type, pointer_buttons, reverse, ctrl, shift, button, pointer_over_canvas, update_helper_layer_on_pointermove_active, pointers, default_palette
    
    8 const { get_tool_by_id, make_monochrome_palette, make_history_node, default_palette, make_canvas, TOOL_PENCIL } = window;
      ~~~~~
    
      src/globals.d.ts:12:1
        12 declare const libtess: any;
           ~~~~~~~
        Conflicts are in this file.
    
    
    Found 1 error in src/app-state.js:8

The line/column numbers are not important here; it's just pointing to
the first non-comment code in both files.

I don't claim to fully understand this, but I was able to resolve it by
doing a sort of bisection, commenting out parts of the file to see how
it affected TypeScript's output.

It turns out that the global declarations for app-state.js are not
needed, and that it doesn't like me importing things from the global
scope into the... global scope. Well that kind of makes sense, when I
say it. But I was aliasing things by destructuring `window`; it's more
verbose now, without that. But not terribly so.

Apparently typescript can find the declarations in `app-state.js` when
type checking other files, even though it's not an ambient declarations
file.

Anyways, the project is now fully lint-free.
No typechecker, linter, or spellchecker errors.