Handle `undefined` defensively
Previously, there was a typo in the name `defaultFileName` (it was
spelled as `defaultName`), leading to an error, fixed in fc59a369fd306b.
I wanted to follow up and see why it gave an error, other than the typo.
With type checking, this sort of problem shouldn't recur, really,
but it "couldn't hurt" to protect against it
(except that it adds complexity of course...)