Fix collapsed white space in certain labels
Previously I used a `display_hotkey` function which was similar to
`AccessKeys.toHTML` but it included a `<span style="white-space: pre">`
wrapper. The wrapper span (although perhaps not the style attribute) is
important to avoid white space collapsing.
For example, in the Flip/Rotate dialog, the "Flip &vertical" label was
converted to 'Flip <span class="menu-hotkey">v</span>ertical', and since
the space between the words was next to an element, it was ignored,
within the context of the `<label>` element parent.