Fix misnomer id
When adding a button to view the project news to the about window in
https://github.com/1j01/jspaint/commit/5d7d135ed895b13
it was visually grouped with the update checking status area.

When I moved the button to a different place in the dialog in
https://github.com/1j01/jspaint/commit/61b22adbc631d03
I neglected to change the id for the container,
which was named in reference to its contents.

I'm not sure the container serves any purpose anymore...