Separate classes and events used by Eye Gaze Mode for UI enlargement
I'm preparing for splitting up the Eye Gaze Mode feature into
several discrete features.

This commit maintains the UI of toggling "Eye Gaze Mode" from the menu,
but changes it internally to use two separate CSS classes, and two
separate toggle events.

This would be a lot to change while also splitting the actual feature
up, so it's better to do it as a refactor first.

I've skipped changing classes for the floating undo and eye gaze pause
buttons in the themes as I'm not sure about that at this point.