Toggle auto-scaling without recreating components
When you toggled "Enlarge UI", previously it recreated the toolbox and
color box components in order to enable/disable code that runs on an
interval to re-scale the components when Enlarge UI is enabled, and to
(implicitly) reset the transformations when disabled.

Now the components handle changes to this setting during their lifetime,
and the relevant code is all in one place.